3.专业教育/计算机科学与技术/专业必修/计算机网络/socket编程/network-pj-chatroom-master/requirements.txt
Pillow==3.4.2
pycrypto==2.6.1